+++ src/api/java/mekanism/api/fluid/IMekanismFluidHandler.java
@@ -8,7 +8,7 @@ import mekanism.api.AutomationType;
 import mekanism.api.IContentsListener;
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.core.Direction;
-import net.minecraftforge.fluids.FluidStack;
+import io.github.fabricators_of_create.porting_lib.transfer.fluid.FluidStack;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
@@ -74,7 +74,7 @@ public interface IMekanismFluidHandler extends ISidedFluidHandler, IContentsList
     }
 
     @Override
-    default int getTankCapacity(int tank, @Nullable Direction side) {
+    default long getTankCapacity(int tank, @Nullable Direction side) {
         IExtendedFluidTank fluidTank = getFluidTank(tank, side);
         return fluidTank == null ? 0 : fluidTank.getCapacity();
     }
@@ -92,7 +92,7 @@ public interface IMekanismFluidHandler extends ISidedFluidHandler, IContentsList
     }
 
     @Override
-    default FluidStack extractFluid(int tank, int amount, @Nullable Direction side, Action action) {
+    default FluidStack extractFluid(int tank, long amount, @Nullable Direction side, Action action) {
         IExtendedFluidTank fluidTank = getFluidTank(tank, side);
         return fluidTank == null ? FluidStack.EMPTY : fluidTank.extract(amount, action, side == null ? AutomationType.INTERNAL : AutomationType.EXTERNAL);
     }
