+++ src/api/java/mekanism/api/recipes/inputs/InputHelper.java
@@ -3,6 +3,8 @@ package mekanism.api.recipes.inputs;
 import java.util.Objects;
 import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
+
+import io.github.fabricators_of_create.porting_lib.transfer.fluid.FluidStack;
 import mekanism.api.Action;
 import mekanism.api.MekanismAPI;
 import mekanism.api.annotations.NonNull;
@@ -16,7 +18,6 @@ import mekanism.api.recipes.cache.CachedRecipe.OperationTracker.RecipeError;
 import mekanism.api.recipes.ingredients.InputIngredient;
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.fluids.FluidStack;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
@@ -151,7 +152,7 @@ public class InputHelper {
                 }
                 FluidStack inputFluid = getInput();
                 if (!inputFluid.isEmpty()) {
-                    int amount = recipeInput.getAmount() * operations;
+                    int amount = (int) (recipeInput.getAmount() * operations);
                     logMismatchedStackSize(tank.shrinkStack(amount, Action.EXECUTE), amount);
                 }
             }
@@ -165,7 +166,7 @@ public class InputHelper {
                     // where we may want to allow not having the input be required for recipe matching
                     if (!recipeInput.isEmpty()) {
                         //TODO: Simulate the drain?
-                        int operations = getInput().getAmount() / (recipeInput.getAmount() * usageMultiplier);
+                        int operations = (int) (getInput().getAmount() / (recipeInput.getAmount() * usageMultiplier));
                         if (operations > 0) {
                             tracker.updateOperations(operations);
                             return;
